+++ config.js
@@ -9,30 +9,61 @@ System.config({
 
 System.config({
   "map": {
-    "aurelia-binding": "github:aurelia/binding@0.4.0",
-    "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.5.0",
-    "aurelia-loader": "github:aurelia/loader@0.4.0",
-    "aurelia-logging": "github:aurelia/logging@0.2.6",
-    "aurelia-metadata": "github:aurelia/metadata@0.3.4",
-    "aurelia-path": "github:aurelia/path@0.4.6",
-    "aurelia-task-queue": "github:aurelia/task-queue@0.2.5",
-    "aurelia-templating": "github:aurelia/templating@0.9.0",
+    "aurelia-binding": "github:aurelia/binding@0.5.0",
+    "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.6.0",
+    "aurelia-loader": "github:aurelia/loader@0.5.0",
+    "aurelia-logging": "github:aurelia/logging@0.3.0",
+    "aurelia-metadata": "github:aurelia/metadata@0.4.0",
+    "aurelia-path": "github:aurelia/path@0.5.0",
+    "aurelia-task-queue": "github:aurelia/task-queue@0.3.0",
+    "aurelia-templating": "github:aurelia/templating@0.10.0",
     "core-js": "github:zloirock/core-js@0.8.1",
     "github:aurelia/binding@0.4.0": {
       "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.5.0",
       "aurelia-metadata": "github:aurelia/metadata@0.3.4",
       "aurelia-task-queue": "github:aurelia/task-queue@0.2.5"
     },
+    "github:aurelia/binding@0.5.0": {
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.6.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.4.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.3.0",
+      "core-js": "github:zloirock/core-js@0.8.1"
+    },
     "github:aurelia/dependency-injection@0.5.0": {
       "aurelia-logging": "github:aurelia/logging@0.2.6",
       "aurelia-metadata": "github:aurelia/metadata@0.3.4",
       "core-js": "npm:core-js@0.4.10"
+    },
+    "github:aurelia/dependency-injection@0.6.0": {
+      "aurelia-logging": "github:aurelia/logging@0.3.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.4.0",
+      "core-js": "github:zloirock/core-js@0.8.1"
     },
     "github:aurelia/loader@0.4.0": {
       "aurelia-html-template-element": "github:aurelia/html-template-element@0.1.3",
       "aurelia-path": "github:aurelia/path@0.4.6",
       "core-js": "npm:core-js@0.4.10",
       "webcomponentsjs": "github:webcomponents/webcomponentsjs@0.5.5"
+    },
+    "github:aurelia/loader@0.5.0": {
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.2.0",
+      "aurelia-path": "github:aurelia/path@0.5.0",
+      "core-js": "github:zloirock/core-js@0.8.1",
+      "webcomponentsjs": "github:webcomponents/webcomponentsjs@0.5.5"
+    },
+    "github:aurelia/metadata@0.4.0": {
+      "core-js": "github:zloirock/core-js@0.8.1"
+    },
+    "github:aurelia/templating@0.10.0": {
+      "aurelia-binding": "github:aurelia/binding@0.5.0",
+      "aurelia-dependency-injection": "github:aurelia/dependency-injection@0.6.0",
+      "aurelia-html-template-element": "github:aurelia/html-template-element@0.2.0",
+      "aurelia-loader": "github:aurelia/loader@0.5.0",
+      "aurelia-logging": "github:aurelia/logging@0.3.0",
+      "aurelia-metadata": "github:aurelia/metadata@0.4.0",
+      "aurelia-path": "github:aurelia/path@0.5.0",
+      "aurelia-task-queue": "github:aurelia/task-queue@0.3.0",
+      "core-js": "github:zloirock/core-js@0.8.1"
     },
     "github:aurelia/templating@0.9.0": {
       "aurelia-binding": "github:aurelia/binding@0.4.0",